Создан модуль Vacancies + множественный рефакторинг
