+++ .eslintrc.js
@@ -16,5 +16,9 @@ module.exports = {
 		sourceType: 'module',
 	},
 	plugins: ['react', 'prettier'],
-	rules: {},
+	rules: {
+		'react/jsx-props-no-spreading': 'off',
+		'import/no-unresolved': 'off',
+		'react/prop-types': 'off',
+	},
 }
