--- src/components/app/index.js
@@ -1,5 +1,0 @@
-import app from './app';
-
-import './app.css';
-
-export default app;
