+++ src/components/form/index.js
@@ -1,8 +1,4 @@
 import Form from './Form'
-import FormItem from './FormItem'
 import './form.css'
 
-export {
-  Form,
-  FormItem
-}
+export default Form
