+++ src/components/index.js
@@ -9,4 +9,6 @@ export { default as Spinner } from './spinner'
 export { default as Checkbox } from './checkbox'
 export { default as Row } from './row'
 export { default as Col } from './col'
-export { Form, FormItem } from './form'
+export { default as Form } from './form'
+export { default as ContextMenu } from './context-menu'
+export { default as Record } from './record'
