+++ src/containers/index.js
@@ -1,11 +1,7 @@
-import VacansyListContainer from './vacancy-list-container';
 import DocumentTableContainer from './document-table-container';
 import SalaryTableContainer from './salary-table-container';
-import VacansyContainer from './vacancy-container';
 
 export {
-  VacansyListContainer,
   DocumentTableContainer,
   SalaryTableContainer,
-  VacansyContainer,
 };
