--- src/containers/vacancy-container/index.js
@@ -1,3 +1,0 @@
-import VacansyContainer from './vacancy-container';
-
-export default VacansyContainer;
