+++ src/modules/vacancies/components/candidate-list-item/candidate-list-item.css
