+++ src/modules/vacancies/components/candidate-list-item/index.js
