+++ src/modules/vacancies/components/index.js
@@ -1,0 +1,11 @@
+import AddVacancyForm from './add-vacancy-form'
+import CandidateList from './candidate-list'
+import VacancyListItem from './vacancy-list-item'
+import CandidateListItem from './candidate-list-item'
+
+export {
+  AddVacancyForm,
+  CandidateList,
+  VacancyListItem,
+  CandidateListItem
+}
