+++ src/modules/vacancies/components/vacancy-list-item/index.js
