+++ src/modules/vacancies/components/vacancy-list-item/vacancy-list-item.css
