+++ src/modules/vacancies/containers/index.js
@@ -1,0 +1,7 @@
+import VacancyListContainer from './vacancy-list-container'
+import VacancyContainer from './vacancy-container'
+
+export {
+  VacancyListContainer,
+  VacancyContainer
+}
