+++ src/modules/vacancies/index.js
@@ -1,0 +1,6 @@
+import { VacancyListContainer, VacancyContainer } from './containers'
+
+export {
+  VacancyListContainer,
+  VacancyContainer
+}
